requirements.txt
annotated-types==0.7.0
anyio==4.6.0
CacheControl==0.14.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==43.0.1
decorator==4.4.2
elevenlabs==1.9.0
fastapi==0.115.2
firebase-admin==6.5.0
google-ai-generativelanguage==0.6.10
google-api-core==2.21.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-generativeai==0.8.3
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
grpcio==1.66.2
grpcio-status==1.66.2
h11==0.14.0
httpcore==1.0.6
httplib2==0.22.0
httpx==0.27.2
idna==3.10
imageio==2.35.1
imageio-ffmpeg==0.5.1
moviepy==1.0.3
msgpack==1.1.0
nest-asyncio==1.6.0
numpy==2.1.2
pillow==10.4.0
proglog==0.1.10
proto-plus==1.24.0
protobuf==5.28.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
PyJWT==2.9.0
pyparsing==3.1.4
requests==2.32.3
rsa==4.9
setuptools==75.1.0
sniffio==1.3.1
starlette==0.39.2
tqdm==4.66.5
twelvelabs==0.2.8
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.31.1
websockets==13.1
wheel==0.44.0